+++ src/finn/benchmarking/bench_base.py
@@ -412,6 +412,7 @@ class bench:
         # cfg.stitched_ip_gen_dcp = False # only needed for further manual integration
         cfg.save_intermediate_models = True  # Save the intermediate model graphs
         cfg.verify_save_full_context = True  # Output full context dump for verification steps
+        cfg.verify_nodewise_report = True  # Per-node deviation report for simulation steps
         cfg.enable_instrumentation = True
         # cfg.default_swg_exception